+++ guppylang/compiler/expr_compiler.py
@@ -4,7 +4,8 @@ from collections.abc import Iterator
 from contextlib import contextmanager
 from typing import Any, TypeGuard, TypeVar
 
-from hugr.serialization import ops
+from hugr.serialization import ops, tys
+from typing_extensions import assert_never
 
 from guppylang.ast_util import AstVisitor, get_type, with_loc, with_type
 from guppylang.cfg.builder import tmp_vars
@@ -31,9 +32,13 @@ from guppylang.nodes import (
     TensorCall,
     TypeApply,
 )
-from guppylang.tys.arg import ConstArg, TypeArg
-from guppylang.tys.builtin import bool_type, get_element_type, is_list_type
-from guppylang.tys.const import ConstValue
+from guppylang.tys.builtin import (
+    bool_type,
+    get_element_type,
+    is_bool_type,
+    is_list_type,
+)
+from guppylang.tys.const import BoundConstVar, ConstValue, ExistentialConstVar
 from guppylang.tys.subst import Inst
 from guppylang.tys.ty import (
     BoundTypeVar,
@@ -304,9 +309,7 @@ class ExprCompiler(CompilerBase, AstVisitor[OutPortV]):
         # since it is not implemented via a dunder method
         if isinstance(node.op, ast.Not):
             arg = self.visit(node.operand)
-            op = ops.CustomOp(
-                extension="logic", op_name="Not", args=[], parent=UNDEFINED
-            )
+            op = ops.CustomOp(extension="logic", name="Not", args=[], parent=UNDEFINED)
             return self.graph.add_node(ops.OpType(op), inputs=[arg]).add_out_port(
                 bool_type()
             )
@@ -319,14 +322,50 @@ class ExprCompiler(CompilerBase, AstVisitor[OutPortV]):
         return unpack.out_port(node.struct_ty.fields.index(node.field))
 
     def visit_ResultExpr(self, node: ResultExpr) -> OutPortV:
-        type_args = [
-            TypeArg(node.ty),
-            ConstArg(ConstValue(value=node.tag, ty=NumericType(NumericType.Kind.Nat))),
-        ]
+        extra_args = []
+        if isinstance(node.base_ty, NumericType):
+            match node.base_ty.kind:
+                case NumericType.Kind.Nat:
+                    base_name = "uint"
+                    extra_args = [
+                        tys.TypeArg(tys.BoundedNatArg(n=NumericType.INT_WIDTH))
+                    ]
+                case NumericType.Kind.Int:
+                    base_name = "int"
+                    extra_args = [
+                        tys.TypeArg(tys.BoundedNatArg(n=NumericType.INT_WIDTH))
+                    ]
+                case NumericType.Kind.Float:
+                    base_name = "f64"
+                case kind:
+                    assert_never(kind)
+        else:
+            # The only other valid base type is bool
+            assert is_bool_type(node.base_ty)
+            base_name = "bool"
+        if node.array_len is not None:
+            op_name = f"result_array_{base_name}"
+            match node.array_len:
+                case ConstValue(value=value):
+                    assert isinstance(value, int)
+                    extra_args = [tys.TypeArg(tys.BoundedNatArg(n=value)), *extra_args]
+                case BoundConstVar():
+                    # TODO: We need to handle this once we allow function definitions
+                    #  that are generic over array lengths
+                    raise NotImplementedError
+                case ExistentialConstVar() as var:
+                    raise InternalGuppyError(
+                        f"Unsolved existential variable during Hugr lowering: {var}"
+                    )
+                case c:
+                    assert_never(c)
+        else:
+            op_name = f"result_{base_name}"
+        args = [tys.TypeArg(tys.StringArg(arg=node.tag)), *extra_args]
         op = ops.CustomOp(
-            extension="tket2.results",
-            op_name="Result",
-            args=[arg.to_hugr() for arg in type_args],
+            extension="tket2.result",
+            name=op_name,
+            args=args,
             parent=UNDEFINED,
         )
         self.graph.add_node(ops.OpType(op), inputs=[self.visit(node.value)])