+++ guppylang/prelude/_internal.py
@@ -22,9 +22,15 @@ from guppylang.definition.value import CallableDef
 from guppylang.error import GuppyError, GuppyTypeError, InternalGuppyError
 from guppylang.hugr_builder.hugr import UNDEFINED, OutPortV
 from guppylang.nodes import GlobalCall, ResultExpr
-from guppylang.tys.arg import ConstArg
-from guppylang.tys.builtin import bool_type, int_type, list_type
-from guppylang.tys.const import ConstValue
+from guppylang.tys.arg import ConstArg, TypeArg
+from guppylang.tys.builtin import (
+    bool_type,
+    int_type,
+    is_array_type,
+    is_bool_type,
+    list_type,
+)
+from guppylang.tys.const import Const, ConstValue
 from guppylang.tys.subst import Inst, Subst
 from guppylang.tys.ty import (
     FunctionType,
@@ -101,9 +107,7 @@ def list_value(v: list[ops.Value], ty: Type) -> ops.Value:
 def logic_op(op_name: str, args: list[tys.TypeArg] | None = None) -> ops.OpType:
     """Utility method to create Hugr logic ops."""
     return ops.OpType(
-        ops.CustomOp(
-            extension="logic", op_name=op_name, args=args or [], parent=UNDEFINED
-        )
+        ops.CustomOp(extension="logic", name=op_name, args=args or [], parent=UNDEFINED)
     )
 
 
@@ -117,14 +121,14 @@ def int_op(
     if args is None:
         args = num_params * [tys.TypeArg(tys.BoundedNatArg(n=NumericType.INT_WIDTH))]
     return ops.OpType(
-        ops.CustomOp(extension=ext, op_name=op_name, args=args, parent=UNDEFINED)
+        ops.CustomOp(extension=ext, name=op_name, args=args, parent=UNDEFINED)
     )
 
 
 def float_op(op_name: str, ext: str = "arithmetic.float") -> ops.OpType:
     """Utility method to create Hugr integer arithmetic ops."""
     return ops.OpType(
-        ops.CustomOp(extension=ext, op_name=op_name, args=[], parent=UNDEFINED)
+        ops.CustomOp(extension=ext, name=op_name, args=[], parent=UNDEFINED)
     )
 
 
@@ -279,27 +283,43 @@ class ArrayLenChecker(CustomCallChecker):
 
 
 class ResultChecker(CustomCallChecker):
-    """Call checker for the `result` function.
-
-    This is a temporary hack until we have implemented the proper results mechanism.
-    """
+    """Call checker for the `result` function."""
 
     def synthesize(self, args: list[ast.expr]) -> tuple[ast.expr, Type]:
         check_num_args(2, len(args), self.node)
         [tag, value] = args
-        if not isinstance(tag, ast.Constant) or not isinstance(tag.value, int):
-            raise GuppyTypeError("Expected an int literal", tag)
+        if not isinstance(tag, ast.Constant) or not isinstance(tag.value, str):
+            raise GuppyTypeError("Expected a string literal", tag)
         value, ty = ExprSynthesizer(self.ctx).synthesize(value)
-        if ty.linear:
-            raise GuppyTypeError(
-                f"Cannot use value with linear type `{ty}` as a result", value
-            )
-        return with_loc(self.node, ResultExpr(value, ty, tag.value)), NoneType()
+        # We only allow numeric values or vectors of numeric values
+        err = (
+            f"Expression of type `{ty}` is not a valid result. Only numeric values or "
+            "arrays thereof are allowed."
+        )
+        if self._is_numeric_or_bool_type(ty):
+            base_ty = ty
+            array_len: Const | None = None
+        elif is_array_type(ty):
+            [ty_arg, len_arg] = ty.args
+            assert isinstance(ty_arg, TypeArg)
+            assert isinstance(len_arg, ConstArg)
+            if not self._is_numeric_or_bool_type(ty_arg.ty):
+                raise GuppyError(err, value)
+            base_ty = ty_arg.ty
+            array_len = len_arg.const
+        else:
+            raise GuppyError(err, value)
+        node = ResultExpr(value, base_ty, array_len, tag.value)
+        return with_loc(self.node, node), NoneType()
 
     def check(self, args: list[ast.expr], ty: Type) -> tuple[ast.expr, Subst]:
         expr, res_ty = self.synthesize(args)
         subst, _ = check_type_against(res_ty, ty, self.node)
         return expr, subst
+
+    @staticmethod
+    def _is_numeric_or_bool_type(ty: Type) -> bool:
+        return isinstance(ty, NumericType) or is_bool_type(ty)
 
 
 class NatTruedivCompiler(CustomCallCompiler):