+++ tests/error/linear_errors/while_unused.py
@@ -1,7 +1,6 @@
 import guppylang.prelude.quantum as quantum
 from guppylang.decorator import guppy
 from guppylang.module import GuppyModule
-from guppylang.prelude.builtins import result
 from guppylang.prelude.quantum import qubit
 
 
@@ -10,8 +9,12 @@ module.load(quantum)
 
 
 @guppy(module)
-def foo(q: qubit) -> None:
-    result("foo", q)
+def test(n: int) -> None:
+    q = qubit()
+    i = 0
+    while i < n:
+        q = h(q)
+        i += 1
 
 
 module.compile()