+++ tests/error/misc_errors/result_array_not_numeric.py
@@ -1,0 +1,7 @@
+from guppylang.prelude.builtins import result, array
+from tests.util import compile_guppy
+
+
+@compile_guppy
+def foo(x: array[tuple[int, bool], 42]) -> None:
+    result("foo", x)